Filled in program
Revieved function heads and I filled in the bodies. The two functions
create a matrix like object. makeCacheMatrix initalizes the object and
defines its methods for interacting with it. cacheSolve is a function
call to solve the inverse of a matrix object. It calculates the inverse
ONCE and stores is at cache. If the function is called again the inverse
will not be solved but rather looked up in memory